app/requirements.txt
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ebaysdk==2.2.0
idna==3.8
lxml==5.3.0
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.2
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.2
wheel==0.42.0
gunicorn
whitenoise
